fixed the homepage...

+++ golang/Server/psql/homepage.go
@@ -36,7 +36,7 @@ func GetHomePageHotnewDate() (date [5]HomePageNews) {
 }
 
 //return homepage aritical data
-func GetHPADate(index int64) (date []ArticlaBox) {
+func GetHPADate(index int) (date []ArticlaBox) {
 	command := "select imgurl, linkurl, brif, date from homepageartical order by id asc offset $1 limit 10"
 	rows, err := db.Query(command, index)
 	var temp ArticlaBox
